Added Game State Engine, Fixed Height Tracking, Added End Game UI Panel Base
Added the countdown, engine states, fixed the scoring for the height tracking (it was tracking the highest it was at all not highest height at the end like it was supposed to). I added the baseline panel for the end game including the high score tracking, final score, and restart button to quickly go again.